refactor(pages): edit sections structure (#45)
Rename software constraints with system components.
Move system components from home page to Arch Linux index.
Update the hardware requirements.
Update the project description in the homepage.
Move the caution section to the top of the page in the homepage.
Remove the toc in the bottom of the page in the Arch Linux index.